Use time.perf_counter when python version >= 3.3
[time.clock()](https://docs.python.org/3.7/library/time.html#time.clock) was deprecated in python 3.3
"Deprecated since version 3.3: The behaviour of this function depends on the platform: use perf_counter() or process_time() instead, depending on your requirements, to have a well defined behaviour."

[time.perf_counter()](https://docs.python.org/3.7/library/time.html#time.perf_counter) was introduced in python 3.3
"Return the value (in fractional seconds) of a performance counter, i.e. a clock with the highest available resolution to measure a short duration. It does include time elapsed during sleep and is system-wide. The reference point of the returned value is undefined, so that only the difference between the results of consecutive calls is valid.

New in version 3.3."